# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecommerce Ordering Payment API",
        "description": "REST API Documentation for the Ecommerce Ordering Payment System.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products",
                "description": "--------------------------------------------------------------------------\nGet All Products\n--------------------------------------------------------------------------\nRetrieve all available products.",
                "operationId": "1ce27f90dd8c39cea9520bf45d58c7c2",
                "responses": {
                    "200": {
                        "description": "Products retrieved successfully."
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create product",
                "description": "--------------------------------------------------------------------------\nCreate Product\n--------------------------------------------------------------------------\nCreate a new product. (Admin Only)",
                "operationId": "acaa0d3fdc74fd4b534e70fc016cf7f1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "name",
                                    "sku",
                                    "price",
                                    "stock",
                                    "status"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "iPhone 16 Pro"
                                    },
                                    "sku": {
                                        "type": "string",
                                        "example": "IP16PRO-256-BLK"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Apple iPhone 16 Pro 256GB Black"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1499.99
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Forbidden. Admin access required."
                    },
                    "422": {
                        "description": "Validation failed."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/products/{product}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product by ID",
                "description": "--------------------------------------------------------------------------\nGet Product\n--------------------------------------------------------------------------\nRetrieve a product by its ID.",
                "operationId": "9d6564b9647b6b46ca1d3eb86c51643e",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product retrieved successfully."
                    },
                    "404": {
                        "description": "Product not found."
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update product",
                "description": "--------------------------------------------------------------------------\nUpdate Product\n--------------------------------------------------------------------------\nUpdate an existing product. (Admin Only)",
                "operationId": "10db88cc185f628118bfa9a84d1401bb",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "iPhone 16 Pro Max"
                                    },
                                    "sku": {
                                        "type": "string",
                                        "example": "IP16PM-512-BLK"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated Apple iPhone 16 Pro Max 512GB"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1699.99
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 40
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Forbidden. Admin access required."
                    },
                    "404": {
                        "description": "Product not found."
                    },
                    "422": {
                        "description": "Validation failed."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete product",
                "description": "--------------------------------------------------------------------------\nDelete Product\n--------------------------------------------------------------------------\nDelete an existing product. (Admin Only)",
                "operationId": "dc780252f855e97451083ce643a83174",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Forbidden. Admin access required."
                    },
                    "404": {
                        "description": "Product not found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum Bearer Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}